+++ CHANGELOG.md
@@ -5,12 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.0] - 2019-10-16
 ### Added
 - addCatalog direct call
 - addParentCatalog direct call
 - updateCatalog direct call
 - getCatalog
 - getCatalogs
+- webhook services added
+- webhook create, get, getList, webhook event DTOs
+- webhook unit and integration tests
 - SimpleFulfillmentOrder - add property fulfillmentOrderAddress
 - Order - add property updateDate
 - LineItem - add property options
@@ -147,7 +152,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - createCategory event & direct call
 - deleteCategory event & direct call
 
-[Unreleased]: https://github.com/shopgate/connect-integration-sdk/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/shopgate/connect-integration-sdk/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/shopgate/connect-integration-sdk/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/shopgate/connect-integration-sdk/compare/v0.9.0...v1.0.0
 [0.9.0]: https://github.com/shopgate/connect-integration-sdk/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/shopgate/connect-integration-sdk/compare/v0.7.0...v0.8.0